Feature/#160 resetting a password

+++ src/app/api/supabase/service.ts
@@ -8,7 +8,7 @@ import { supabase } from '@/app/api/supabase/client';
 import { InsertNewPlansType, PlansType } from '@/types/plans';
 import { CONTACTS, PLANS, USERS } from '@/constants/supabaseTable';
 import { User } from '@supabase/supabase-js';
-import { REDIRECT_TO } from '@/constants/redirecturl';
+import { REDIRECT_TO, REDIRECT_TO_CHANGE_PASSWORD, REDIRECT_TO_FINISH_SIGNUP } from '@/constants/redirecturl';
 
 export const getContacts = async (userId: string): Promise<ContactItemType[]> => {
   try {
@@ -61,7 +61,7 @@ export const signUpUser = async (value: { email: string; password: string; nickn
   const { data, error } = await supabase.auth.signUp({
     email,
     password,
-    options: { data: { nickname } },
+    options: { data: { nickname }, emailRedirectTo: REDIRECT_TO_FINISH_SIGNUP },
   });
 
   return { data, error };
@@ -324,27 +324,19 @@ export const getSelectPlan = async (plan_id: string) => {
 };
 
 // contacts 데이터 삭제
-export const mutateDeleteContacts = async (
-  userId: string,
-  contactsId: string
-): Promise<void> => {
+export const mutateDeleteContacts = async (userId: string, contactsId: string): Promise<void> => {
   try {
-    const { error } = await supabase
-      .from(CONTACTS)
-      .delete()
-      .eq('contacts_id', contactsId)
-      .eq('user_id', userId)   // userId 조건 추가
+    const { error } = await supabase.from(CONTACTS).delete().eq('contacts_id', contactsId).eq('user_id', userId); // userId 조건 추가
 
     if (error) {
-      console.error('연락처 삭제 중 오류가 발생했습니다:', error)
-      throw new Error('연락처 삭제 중 문제가 발생했습니다.')
+      console.error('연락처 삭제 중 오류가 발생했습니다:', error);
+      throw new Error('연락처 삭제 중 문제가 발생했습니다.');
     }
-
   } catch (error) {
-    console.error('연락처 삭제 요청 실패:', error)
-    throw new Error('연락처 삭제 중 문제가 발생했습니다. 다시 시도해주세요.')
+    console.error('연락처 삭제 요청 실패:', error);
+    throw new Error('연락처 삭제 중 문제가 발생했습니다. 다시 시도해주세요.');
   }
-}
+};
 
 // plans 데이터 삭제
 export const mutateDeletePlan = async (planId: string): Promise<void> => {
@@ -402,3 +394,24 @@ export const fetchRegularContactsInfinite = async (
   return { contacts: data, nextPage };
 };
 
+export const changePassword = async (newPassword: string) => {
+  const { data, error } = await supabase.auth.updateUser({ password: newPassword });
+  if (error) {
+    console.error('비밀번호 재설정 중 문제가 발생했습니다.', error);
+    throw new Error('비밀번호 재설정 중 문제가 발생했습니다. 다시 시도해주세요.');
+  }
+
+  return data;
+};
+
+export const sendPasswordResetEmail = async (email: string) => {
+  const { data, error } = await supabase.auth.resetPasswordForEmail(email, {
+    redirectTo: REDIRECT_TO_CHANGE_PASSWORD,
+  });
+  if (error) {
+    console.error('비밀번호 재설정 이메일 전송 중 문제가 발생했습니다.', error);
+    throw new Error('비밀번호 재설정 이메일 전송 중 문제가 발생했습니다. 다시 시도해주세요.');
+  }
+
+  return data;
+};
